toolkit/components/url-classifier/tests/mochitest/test_classifier_worker.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the URI Classifier</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script class="testbody" type="text/javascript">

// Add some URLs to the malware database.
var testData = "example.com/tests/toolkit/components/url-classifier/tests/mochitest/evilWorker.js";
var testUpdate =
  "n:1000\ni:test-malware-simple\nad:550\n" +
  "a:550:32:" + testData.length + "\n" +
  testData;

testData = "example.com/tests/toolkit/components/url-classifier/tests/mochitest/unwantedWorker.js";
testUpdate +=
  "n:1000\ni:test-unwanted-simple\nad:550\n" +
  "a:550:32:" + testData.length + "\n" +
  testData;

function loadTestFrame() {
  document.getElementById("testFrame").src =
    "http://example.com/tests/toolkit/components/url-classifier/tests/mochitest/workerFrame.html";
}

function onmessage(event)
{
  var pieces = event.data.split(':');
  if (pieces[0] == "finish") {
    SimpleTest.finish();
    return;
  }

  is(pieces[0], "success", pieces[1]);
}

const CLASSIFIER_COMMON_URL = SimpleTest.getTestFileURL("classifierCommon.js");
let classifierCommonScript = SpecialPowers.loadChromeScript(CLASSIFIER_COMMON_URL);

classifierCommonScript.addMessageListener("loadTestFrame", () => {
  SpecialPowers.pushPrefEnv(
    {"set" : [["browser.safebrowsing.malware.enabled", true]]},
    loadTestFrame);
});
classifierCommonScript.addMessageListener("updateError", ({ errorCode }) => {
  ok(false, "Couldn't update classifier. Error code: " + errorCode);
  // Abort test.
  SimpleTest.finish();
});

SpecialPowers.pushPrefEnv(
  {"set" : [["urlclassifier.malwareTable", "test-malware-simple,test-unwanted-simple"],
            ["urlclassifier.phishTable", "test-phish-simple"]]},
  function() {
    classifierCommonScript.sendAsyncMessage("doUpdate", { testUpdate });
  });

window.addEventListener("message", onmessage, false);

SimpleTest.waitForExplicitFinish();

</script>

</pre>
<iframe id="testFrame" onload=""></iframe>
</body>
</html>
